try using simple namespace instead of receiver
