+++ src/jwk/resolver.rs
@@ -18,7 +18,7 @@ use serde_json::json;
 use super::DidJwk;
 use crate::document::{CreateOptions, MethodType};
 use crate::error::Error;
-use crate::resolution::{ContentType, Metadata, Options, Resolved};
+use crate::operation::resolve::{ContentType, Metadata, Options, Resolved};
 use crate::{DidOperator, DidResolver, KeyPurpose, PublicKeyJwk};
 
 static DID_REGEX: LazyLock<Regex> =
